pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["factory"]

[tool.hatch.build.targets.wheel.force-include]
"templates" = "factory/templates"
"schemas" = "factory/schemas"
"registry" = "factory/registry"

[project]
name = "agent-factory"
version = "1.0.0"
description = "Generate AI agent repositories through LLM-guided interviews"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "types-PyYAML",
    "types-jsonschema",
]

[project.scripts]
factory = "factory.cli.main:cli"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
